apps/backend/requirements.txt
# Основные зависимости
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.3
pydantic-settings==2.2.1
# Обработка изображений
pillow==10.2.0
piexif==1.1.3
minio==7.2.3
celery==5.3.6
redis==5.0.1
httpx==0.27.0
python-dotenv==1.0.1

# Кеширование
redis==5.0.1
redis-py-cluster==2.1.3

# Тестирование
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
testcontainers==3.7.1

# Утилиты для разработки
black==24.3.0
flake8==7.0.0
isort==5.12.0
mypy==1.5.1

